refactor: reduce migration builder API to a single type.
